Correct cancel event for touchable spans
Summary:
While adding missing props to RCTextView it turned out there is some msleading behaviour with cancel event.

with the previous implementation if we have cancel, we would send it with null span. But in code people are sometimes comparing spans before doing any action, like here:
https://www.internalfb.com/code/fbsource/[0434253bdab9]/fbandroid/java/com/facebook/fds/comment/FDSInternalCommentBubble.kt?lines=139%2C146

so trying in this diff to add proper cancel handling.
we would call cancel with null span if the event happened outside of bounds
we would call cancel with null span if we werent able to find a span that matches the coordinates
and for the rest of the cases we would call cancel with the span that we found

Reviewed By: pentiumao, pasqualeanatriello

Differential Revision: D73914984

fbshipit-source-id: 72533cd6079d756686357dd9d7e12b7468d11fc6