Fix keyboard crash during TextInput measurement
Summary:
I've already tried fixing it in D73921811 but it turns out that some OEMs (Motorola) have custom code that is executed during EditText instantiation and that code accesses InputMethodManager without checkif if it's not null. So my previous diff caused a NPE on those devices. See: https://fburl.com/logview/henk4tk9

This change should address that issue and fix the TextInput measurement crash that happens because of keyboard being hidden.

Reviewed By: astreet

Differential Revision: D74304605

fbshipit-source-id: 2e92b14e715b144c23c6bdc55c6a6943536bc8a3