default to legacy LooperMode
Summary:
D73854866 aims to remove this override from the `robolectric4_test` macro so that the default becomes the non-deprecated PAUSED LooperMode across. But this target times out when the default LooperMode is set to PAUSED.

This diff adds the vm_arg to the target so that it won't start timing out once the override is removed.

More context in this post: https://fb.workplace.com/groups/android.testing.fyi/permalink/2448074072235887/

Reviewed By: cpepin

Differential Revision: D74293898

fbshipit-source-id: 55885d347da5cfc2e0caa70ce5727e3aa0d42447