Add support for clickableSpanExpandedOffset
Summary:
Adding missing functionality that we have in TextSpec.

this was originally added in D3259278

The idea behind it is that you draw the circle around the touch area, and get the selection of spans within that touch area.  You then intersect them.  The Region.op() call will return true if the two intersect (meaning the touch was close), or false if they don't.

Reviewed By: zielinskimz

Differential Revision: D74325713

fbshipit-source-id: 27bc1d157a2fe1778319dbf70366687501e4d19f