Annotate prop `text` as nullable
Summary:
Title.

## General Context

This diff is part of an effort to make all of our Android apps nullsafe. This will not only help reduce NPEs but allow for smooth conversion from Java to Kotlin. If this is your team’s code, we would appreciate it if you could review our efforts as well as make the rest of your codebase nullsafe.

Differential Revision: D74483226

fbshipit-source-id: 39a68459d2f2d8a926e989e37a8a7d74d24eeb6d