Fix crash in TruncateText
Summary: The original way this was implemented made it possible for ellipsisOffset to be less than getLineStart, which caused a crash because of an out of bounds exception. I am adding a check for Math.max(ellipsisOffset, getLineStart), to prevent the out of bounds exception. The code will still function in the same way because the string will be empty in this case and will make noLettersOrNumbers = true.

Reviewed By: AlexBalo

Differential Revision: D74817320

fbshipit-source-id: c6061d07e05934cd4ec9ab2ff59a0ce441db4710