Add RCTextStyleHelper
Summary:
I got a bit ahead with this, but at some point we will delete TextSpec and then we will need a way to do the same as we are doing with text spec but for rctextview.
This is similar class to TextStyleHelper but instead of passing all of the props, we are just directly set them to the rc text style.

Reviewed By: zielinskimz

Differential Revision: D74814921

fbshipit-source-id: af99b76716a3fcc8c44d856026d026d5890a797c