Add QE to use RCTextView instead of TextSpec
Summary:
Previous diff was too big so i decided to split it...
Its not making big difference but this is RCTextView added.
Why it has all of those if else?
I hope that one day this wont be a textspec and then we will have to support xml style loading (class is already here)
so we dont want to override the xml style with default values, so if the value is == to default, we dont do anything, if its different then we are ok with overriding xml style or just setting it.

Reviewed By: zielinskimz

Differential Revision: D74998516

fbshipit-source-id: e4d454df19f3a054edb9b47eb23e2d9bf6f31496