Add Text Primitive and TextSpec wrapper for QE
Summary:
This diff was supposed to add QE, but was so big that I decided to split it.

So this diff is just adding wrapper around TextSpec, changing TextSpec -> TextComponentSpec,
and all of those lines changed are a result of that.
We have lots of tests that are testing against the hierarchy String so when i added the wrapper I had to change the tests too...

I deleted some that didn't make sense and I will re-add them once we remove the wrapper and access RCTextView directly.

Reviewed By: zielinskimz

Differential Revision: D65338516

fbshipit-source-id: 8691d773ce6418c89b8646998ea45a433726fcaa