Set the same break strategy for RCTextView
Summary:
RCTextView was created with different default break strategy.
I think at this point we cannot just change it as it can break the current usage (?)
So the idea is to set it to the default value here where we are redirecting the TextSpec usages.

Then once we get rid of TextSpec, I will have here, after the setter of break strategy, loading of xml style values, and then if the property of the Text has custom break strategy we will set it in the line 238.

so use case:
1) Text("test", breakStrategy= something) ->
i set the default value first, then i set it to something

2) Text("test", style res= R.style.somethingWithBreakStrategy)
i set it to default value, then i load the xml style and set the style from xml then i dont set it in the setter as the value from the prop is equal to default

3)  Text("test",  breakStrategy= something, style res= R.style.somethingWithBreakStrategy)
i set it to default value, then i load the xml style and set it to the strategy from xml, then i set the value from prop.

4) Text("test")
i set it only to the default value

Reviewed By: zielinskimz

Differential Revision: D75132683

fbshipit-source-id: 33981d1fcd951e5e178046c781d6cd45c5c8b1b2