Fix ForceInlineTruncation
Summary: I am now checking forceInlineTruncation before truncate text is called, so that if forceInlineTruncation is enabled, we set the max lines of the text to be the last line of text before the actual max lines value. This allows us to use the existing truncation logic with a new max lines value rather than trying to adjust once the text has been truncated.

Reviewed By: astreet

Differential Revision:
D75547481

Privacy Context Container: L1241033

fbshipit-source-id: 92a01af71f568132604e72a8fc7666a0c8727651