Introduce BinderObserver API
Summary:
This diff introduces a couple of new concepts that will serve as the foundation to make it possible to observe binder execution in a seamless manner.

**BinderType**
This is an enum of either MOUNT or ATTACH binders. Fixed and optional binders are MOUNT based while the attach binders are ATTACH based

**BinderId:**
This is simply a unique identifier for a binder within a render tree. It's made up of the render unit ID, the type and key of the binder.
The type of the binder of a BinderType which is described above
The key of the binder is a BinderKey which provides a 3rd dimension of uniqueness. In fact, for any group of non-fixed binders in a render unit (optional, or attach), they will already have a unique type that can be used to identify them. Fixed binders will specifically use an Index based BinderKey to ensure that they also continue to remain unique within the RenderUnit
Bringing all 3 props together makes it possible to create a unique identity for any binder within its render tree

**BinderObserver:**
This API exposes a way to hook into the execution of binders. This will be used as an entry point with which we can track state reads in the mount phase.

Further work to integrate this API into the render unit will continue in the following diff

Reviewed By: adityasharat

Differential Revision: D74670288

fbshipit-source-id: c54d8757cad3c956a881331ed202ec608b5d0010