Introduce MountContext
Summary:
This diff introduces another new foundational concept

**MountContext:**
This will now serve as a container for mount parameters that can be passed to the render unit. Previously, we explicitly passed the Android context and a sys tracer. However, adding more parameters means that we must thread the change across the entire system which isn't desirable. Having this container makes it easy for such changes to happen without necessarily changing the entire world. For example, we have now been able to seamlessly introduce a BinderObserver parameter

Reviewed By: adityasharat

Differential Revision: D75534555

fbshipit-source-id: e43312e589ec77f03d7ef034e04b740383b900e6