Fix the ClassCastException
Summary:
context: https://fb.workplace.com/groups/flippersupport/permalink/2105908086556505/

The main issue here is `TransientField.CONTENT_FIELD` pass a reflect field and no editor can parse it or random editor is picked up to parse it.

The fixes:
if we know there is no editor to parse it, just return the value of that content.

Reviewed By: LukeDefeo

Differential Revision: D75871900

fbshipit-source-id: c94e62ee84e0bdf3ee82f9a0506ab8054fb87508