use State instead of Cached value
Summary:
There was a support request where meta ai after reconnecting was unable to update the state (and render text that AI is saying)
https://fb.workplace.com/groups/litho.support/permalink/4112144809106856/

This bug is due to the fact that after reconnecting the voice session and triggering the final unmount of the vertical scroll:
https://www.internalfb.com/code/fbsource/[254e1e5eb8a1]/fbandroid/java/com/meta/metaai/voicecard/core/impl/fragment/VoiceCardFragment.kt?lines=805

we were setting the treeState to null.
https://www.internalfb.com/code/fbsource/[d1596e3c32314304fd1f66c256f39a8369454f25]/fbandroid/libraries/components/litho-core/src/main/java/com/facebook/litho/PrimitiveComponentScope.kt?lines=348

After that even though we were recreating the scroll we were still using the null value that we cached here:
```
 val nestedTreeState = useCached(Unit) { NestedLithoTreeState(treeState = TreeState()) }
```
flipping the mc for useStateForCachedValues is fixing the issue, but as this is in QE I would rather land this quick fix and later change it back to cached value once the qe is shipped.

Reviewed By: adityasharat, aashay-gaikwad

Differential Revision: D76603512

fbshipit-source-id: 484ef9a463521e6e1c381ae511468ed0fad68bc3