Fix Primitive TextInput measurement bug
Summary:
There was an issue in the measurement logic of the ExperimentalTextInput. In order to avoid concurrent modifications of mutable spans, we're creating `textForMeasure` state which holds a CharSequence without spans. It's used during measurement.

We're updating the value of `textForMeasure` in onTextChanged callback. However we're only doing it when the number of lines changed. We incorreclty assumed that we need to remeasure only when the number of lines chanes.
In practice, this is not always the case. For example NotesBubble in M4A measures the text manually and sets exact width on the TextInput component on each text change. This means that the TextInput will be remeasured because it gets new SizeConstraints. But `textForMeasure` is only updated when the line count changes, which means that we're remeasuring more often than updating the text and this causes a visual bug.

For example if the `textForMeasure` is has 2 lines, and we remove character by character until it reaches 1 line, it's value will stay the same even if we remove all characters. Then we re-measure the component with smaller widths, but the `textForMeasure` is still having that old "full line" value. This behavior is visible on the first video in the test plan.

The fix is to keep track of not only the text without spans, but also latest size constraints that were available when we updated `textForMeasure` state. Then during the measurement, we check if the size constraints that were visible when `textForMeasure` got updated are the same as the size constraints used to measure the component. If they match then we can safely use the text form `textForMeasure` state. Otherwise we use original text.

I'll be running another QE to confirm if this change doesn't break anything.

Reviewed By: astreet

Differential Revision: D76711809

fbshipit-source-id: c01c6eb6d824adb3881a79919c887a837b82d126