Fix ColorStateList for RCTextView
Summary:
I'm still trying to get to the bottom of this issue.
But it looks like TextSpec is not working correctly as well.

When I switch TextSpec to RCTextView some of the screenshot tests are failing as TextSpec is not correctly showing the ColorStateList values... (so the initial value is wrong in all of the screenshot tests)
RCTextView is doing it correctly but its not changing them.
So this diff is adding proper handling for state changes.

Reviewed By: astreet

Differential Revision: D76818899

fbshipit-source-id: 0ff9be8bcf76c8a4104e6fd1758466504b545747