Use seqNumber for remeasuring TextInput
Summary:
In this diff I'm adding another way of triggering remeasure of the TextInput. It's the same as the Spec version is using and simpler than the current solution. I've added MC and will run a QE to see if it works correctly and if it does then I'll replace the current logic with `textAndConstraintsForMeasure` State.

run_all_fbandroid_tests
run_all_fbandroid_tests_at_diff_time

Reviewed By: astreet

Differential Revision: D76815115

fbshipit-source-id: 5c954d54184f93ca40014a4b7ccf0a98323d3d8e