Correct span click handling
Summary:
During a recent QE, we encountered an issue with the new span click/touch handling approach. A support request from M4A revealed that this approach broke link/mention handling.
After investigating the problem, we found that the default false value for onTouchListener in onTouch was causing us to return false and not receive subsequent events.
To improve test coverage and prevent similar issues in the future, I am open to suggestions on how to better test this functionality.
Since this behavior is triggered by onTouch, simply running onClick does not reproduce the issue.

I have added tests that trigger the action down within the span and handle the click defined on the span differently depending on whether we implement onClick in the span click listener (returning true) or not (returning false).

Reviewed By: zielinskimz

Differential Revision: D76986249

fbshipit-source-id: 1428c8ebbf579cd6eb41e1681502c475d258144b