Call before, mount and after together. (maybe fix T227922576)
Summary:
there was a recent fix of a bug D75218906 -> D75218907
but it looks like we added conditional `beforeMount`
`beforeMount`, `mount` and `onAfterMount` must always be called together or not at all.

So this is fixing something, and maybe this will fix the T227922576.
Im still trying to get a proper local repro with fix and without so will update the diff test plan once it happen.

Reviewed By: adityasharat

Differential Revision: D77028827

fbshipit-source-id: 27b9274a46240f8627248621c10443dde3edcb98