Add an assertion during unmount
Summary:
There is this crash T228928314 and our hypothesis is that there is somehow an extension that causes an unmount which clears the list that is currently iterated on and this way the list gets cleared.

In this diff I'm adding an assertion that will throw if unmount is called during mount It's gated with a MC and I'll run a QE to confirm if that's the case.

Reviewed By: adityasharat

Differential Revision: D77534195

fbshipit-source-id: 6581eddbf22d79066c8ea1b9a3d724d70986eb95