Restore old logic for State equals and hashCode
Summary:
To enable us turn on state read tracking for a quick experiment, we'd temporarily restore the old state identity logic.

Note that there's a codemod running that tries to migrate sites with state object dependency to use their actual values. Once that is complete, we'd be able to revert this back to the expected logic

Reviewed By: adityasharat

Differential Revision: D77530488

fbshipit-source-id: bee736b0f8962fcc943ebc5a6212df64de8c25a2