Add ComponentExtensions for resolving resources
Summary:
As per title. In this diff we're renaming `DimenExtensions` file to `ComponentContextExtensions` and we're adding all APIs available in RenderCore's `Resources.kt` [file](https://www.internalfb.com/code/fbsource/[cb37bd9a61b0]/fbandroid/libraries/rendercore/rendercore/src/main/java/com/facebook/rendercore/Resources.kt?lines=31).

It may seem a bit redundant, but with this we're allowing to call those methods on the ComponentContext. This will simplify KCompoent conversions. Right now when the callsite needs to resolve a resource it has to use the ResourceResolver which is a lot of boilerplate, for example:

```
.contentDescription(checkNotNull(context.resourceResolver.resolveStringRes(R.string.x)))
```

with this change, the above becomes:
```
// kotlin spec
.contentDescription(context.stringRes(R.string.x))

// java spec
.contentDescription(stringRes(context, R.string.title))
```

Ideally, when all Specs are converted, this file can be removed.

Reviewed By: adityasharat

Differential Revision: D78000439

fbshipit-source-id: a428ae1ed4ae9fcd39f125e38120ff39d678b1c1