Add customHostViewBinders as additional binders in createRootHostRenderUnit call
Summary:
While working on drawing api we noticed that when you try to draw on a Column that is the root, we are not seeing anything drawn.
After debugging the issue we were never adding the additional binders (that contains our drawing instructions) so adding the piece that is doing it now.

Reviewed By: adityasharat

Differential Revision: D78266195

fbshipit-source-id: ddd2719467d7cc395526fcb7bc59ec044c94400e