Fix tree ID mismatch when executing effects in CT
Summary:
Fixing a potential tree ID mismatch when executing effects that read state. This can happen for effects from a nested tree that are triggered inline, within an active layout. In such case, the state recorder has an active tree, but the effect is actually running from another (sub)tree, so we think that the state is being read from that (outer)tree, causing unnecessary crashes

This diff fixes the issue by simply ensuring that the effects are executed in the context of their CT. This is done by starting an explicit recorder and passing the current tree ID. We don't actually need to do anything with the states read in this context r

Reviewed By: adityasharat

Differential Revision: D77530490

fbshipit-source-id: 2607975aa550e243a62cfcb0b7251e3f511fd756