Correct the behaviour for RTL language in horizontal scroll
Summary:
While looking into T231423202 where UR was not focusing on the right side of the scroll for arabic language I noticed that there is a bug in scroll implementation.

previously we would set -1 as the unset scroll position, but after calling: toPixels() we would get -3. So the condition was never true:
```
position.x == LAST_SCROLL_POSITION_UNSET
```

which is good as this part of code was wrong.

This diff adds a check for layout direction, and if its RTL then we focus scroll on the right end (previously we would do that for all (?))

and as well putting -1 as the unset scroll position so we can handle those cases properly now.

Reviewed By: zielinskimz

Differential Revision: D78554026

fbshipit-source-id: c074ca7ac527bded0bd93720c260a105ccd1a8fd